Split source lane runtime modules
