+++ src/app/controller/state/runtime/source_lane.rs
@@ -6,6 +6,10 @@ use std::collections::{BTreeSet, HashMap, HashSet};
 use std::path::PathBuf;
 use std::time::Instant;
 
+mod auto_rename_batch;
+mod hydration;
+mod mutations;
+
 /// Source-lane runtime state grouped by hydration, folder projection, and mutations.
 #[derive(Debug, Default)]
 pub(crate) struct SourceLaneRuntimeState {
@@ -24,59 +28,6 @@ pub(crate) struct SourceHydrationRuntime {
     pub(crate) pending_active: Option<PendingSourceHydration>,
     /// Inactive-pane source hydration currently preparing one retained folder snapshot.
     pub(crate) pending_inactive: Option<PendingSourceHydration>,
-}
-
-impl SourceHydrationRuntime {
-    /// Return the pending hydration request for `kind`, when one exists.
-    pub(crate) fn pending(&self, kind: SourceHydrationKind) -> Option<&PendingSourceHydration> {
-        match kind {
-            SourceHydrationKind::ActiveSelection => self.pending_active.as_ref(),
-            SourceHydrationKind::InactivePane => self.pending_inactive.as_ref(),
-        }
-    }
-
-    /// Return the mutable pending hydration request for `kind`, when one exists.
-    pub(crate) fn pending_mut(
-        &mut self,
-        kind: SourceHydrationKind,
-    ) -> Option<&mut PendingSourceHydration> {
-        match kind {
-            SourceHydrationKind::ActiveSelection => self.pending_active.as_mut(),
-            SourceHydrationKind::InactivePane => self.pending_inactive.as_mut(),
-        }
-    }
-
-    /// Store a new pending hydration request in the lane identified by `kind`.
-    pub(crate) fn set_pending(
-        &mut self,
-        kind: SourceHydrationKind,
-        pending: PendingSourceHydration,
-    ) {
-        match kind {
-            SourceHydrationKind::ActiveSelection => self.pending_active = Some(pending),
-            SourceHydrationKind::InactivePane => self.pending_inactive = Some(pending),
-        }
-    }
-
-    /// Clear the pending hydration request for `kind`.
-    pub(crate) fn clear_pending(&mut self, kind: SourceHydrationKind) {
-        match kind {
-            SourceHydrationKind::ActiveSelection => self.pending_active = None,
-            SourceHydrationKind::InactivePane => self.pending_inactive = None,
-        }
-    }
-
-    /// Return the source id currently owning a visible loading state, when any.
-    pub(crate) fn loading_source_id(&self) -> Option<SourceId> {
-        self.pending_active
-            .as_ref()
-            .map(|pending| pending.source_id.clone())
-            .or_else(|| {
-                self.pending_inactive
-                    .as_ref()
-                    .map(|pending| pending.source_id.clone())
-            })
-    }
 }
 
 /// Runtime tracking for pane-scoped folder projection requests.
@@ -118,286 +69,6 @@ pub(crate) struct SourceMutationRuntime {
     active_auto_rename_batch: Option<ActiveAutoRenameBatchState>,
 }
 
-impl SourceMutationRuntime {
-    /// Track one optimistic metadata request and its affected relative paths.
-    pub(crate) fn insert_metadata_mutation(&mut self, pending: PendingMetadataMutation) {
-        let source_id = pending.source_id.clone();
-        for path in &pending.paths {
-            self.pending_metadata_paths
-                .insert((source_id.clone(), path.clone()));
-            if pending.blocks_file_mutation {
-                self.pending_file_blocking_metadata_paths
-                    .insert((source_id.clone(), path.clone()));
-            }
-        }
-        self.pending_metadata_mutations
-            .insert(pending.request_id, pending);
-    }
-
-    #[cfg(test)]
-    /// Return the number of metadata mutation requests currently awaiting completion.
-    pub(crate) fn pending_metadata_count(&self) -> usize {
-        self.pending_metadata_mutations.len()
-    }
-
-    /// Remove one metadata request and clear its tracked pending paths.
-    pub(crate) fn finish_metadata_mutation(
-        &mut self,
-        request_id: u64,
-    ) -> Option<PendingMetadataMutation> {
-        let pending = self.pending_metadata_mutations.remove(&request_id)?;
-        for path in &pending.paths {
-            self.pending_metadata_paths
-                .remove(&(pending.source_id.clone(), path.clone()));
-            if pending.blocks_file_mutation {
-                self.pending_file_blocking_metadata_paths
-                    .remove(&(pending.source_id.clone(), path.clone()));
-            }
-        }
-        Some(pending)
-    }
-
-    /// Return whether one sample path still has an optimistic metadata write in flight.
-    pub(crate) fn metadata_path_pending(
-        &self,
-        source_id: &SourceId,
-        relative_path: &std::path::Path,
-    ) -> bool {
-        self.pending_file_blocking_metadata_paths
-            .contains(&(source_id.clone(), relative_path.to_path_buf()))
-    }
-
-    /// Return whether any optimistic metadata write is pending for `source_id`.
-    pub(crate) fn source_has_pending_metadata(&self, source_id: &SourceId) -> bool {
-        self.pending_metadata_paths
-            .iter()
-            .any(|(pending_source_id, _)| pending_source_id == source_id)
-    }
-
-    /// Extend the source-scoped claim-pause grace window through `until`.
-    pub(crate) fn extend_claim_pause_grace(&mut self, source_id: &SourceId, until: Instant) {
-        let entry = self
-            .claim_pause_grace_until
-            .entry(source_id.clone())
-            .or_insert(until);
-        if *entry < until {
-            *entry = until;
-        }
-    }
-
-    /// Return whether the source-scoped claim-pause grace window is still active.
-    pub(crate) fn claim_pause_grace_active(&mut self, source_id: &SourceId, now: Instant) -> bool {
-        match self.claim_pause_grace_until.get(source_id).copied() {
-            Some(until) if until > now => true,
-            Some(_) => {
-                self.claim_pause_grace_until.remove(source_id);
-                false
-            }
-            None => false,
-        }
-    }
-
-    /// Extend the source-scoped auto-sync suppression window through `until`.
-    pub(crate) fn extend_auto_sync_grace(&mut self, source_id: &SourceId, until: Instant) {
-        let entry = self
-            .auto_sync_grace_until
-            .entry(source_id.clone())
-            .or_insert(until);
-        if *entry < until {
-            *entry = until;
-        }
-    }
-
-    /// Return whether the source-scoped auto-sync suppression window is active.
-    pub(crate) fn auto_sync_grace_active(&mut self, source_id: &SourceId, now: Instant) -> bool {
-        match self.auto_sync_grace_until.get(source_id).copied() {
-            Some(until) if until > now => true,
-            Some(_) => {
-                self.auto_sync_grace_until.remove(source_id);
-                false
-            }
-            None => false,
-        }
-    }
-
-    /// Return whether one source currently owns a background file mutation.
-    pub(crate) fn source_has_pending_file_mutations(&self, source_id: &SourceId) -> bool {
-        self.pending_file_mutation_sources.contains(source_id)
-    }
-
-    /// Mark one source/path batch as owned by a background file mutation.
-    pub(crate) fn begin_file_mutation(
-        &mut self,
-        source_id: &SourceId,
-        paths: impl IntoIterator<Item = PathBuf>,
-    ) -> bool {
-        let source_was_inactive = self.pending_file_mutation_sources.insert(source_id.clone());
-        for path in paths {
-            self.pending_file_mutation_paths
-                .insert((source_id.clone(), path));
-        }
-        source_was_inactive
-    }
-
-    /// Clear one source/path batch from background file-mutation tracking.
-    pub(crate) fn finish_file_mutation(
-        &mut self,
-        source_id: &SourceId,
-        paths: impl IntoIterator<Item = PathBuf>,
-    ) -> bool {
-        for path in paths {
-            self.pending_file_mutation_paths
-                .remove(&(source_id.clone(), path));
-        }
-        let still_has_paths = self
-            .pending_file_mutation_paths
-            .iter()
-            .any(|(pending_source_id, _)| pending_source_id == source_id);
-        if !still_has_paths {
-            return self.pending_file_mutation_sources.remove(source_id);
-        }
-        false
-    }
-
-    /// Mark one browser rename or auto-rename intent as the file-op lane owner.
-    pub(crate) fn begin_browser_rename_intent(&mut self, key: BrowserRenameIntentKey) {
-        self.active_browser_rename_intent = Some(key);
-    }
-
-    /// Drop the active browser rename owner and return any queued follow-up.
-    pub(crate) fn finish_browser_rename_intent(
-        &mut self,
-    ) -> Option<PendingBrowserAutoRenameIntent> {
-        self.active_browser_rename_intent = None;
-        self.active_auto_rename_batch = None;
-        self.queued_browser_auto_rename_intent.take()
-    }
-
-    /// Forget active browser rename ownership when dispatch fails before work starts.
-    pub(crate) fn clear_browser_rename_intent(&mut self) {
-        self.active_browser_rename_intent = None;
-        self.active_auto_rename_batch = None;
-    }
-
-    /// Start tracking one active browser auto-rename batch in requested path order.
-    pub(crate) fn begin_auto_rename_batch(&mut self, source_id: SourceId, paths: Vec<PathBuf>) {
-        self.active_auto_rename_batch = Some(ActiveAutoRenameBatchState::new(source_id, paths));
-    }
-
-    /// Apply one structured auto-rename progress message to the active batch state.
-    pub(crate) fn apply_auto_rename_progress(&mut self, progress: SampleAutoRenameProgress) {
-        if let Some(batch) = self.active_auto_rename_batch.as_mut() {
-            batch.apply_progress(progress);
-        }
-    }
-
-    /// Clear active auto-rename state when the selected source changes.
-    pub(crate) fn clear_auto_rename_batch_for_source_change(
-        &mut self,
-        selected: Option<&SourceId>,
-    ) {
-        if self
-            .active_auto_rename_batch
-            .as_ref()
-            .is_some_and(|batch| Some(&batch.source_id) != selected)
-        {
-            self.active_auto_rename_batch = None;
-        }
-    }
-
-    /// Return a source-scoped immutable snapshot of the active auto-rename batch.
-    pub(crate) fn active_auto_rename_batch_snapshot(
-        &self,
-    ) -> Option<ActiveAutoRenameBatchSnapshot> {
-        self.active_auto_rename_batch
-            .as_ref()
-            .map(ActiveAutoRenameBatchState::snapshot)
-    }
-
-    /// Apply the OPT-135 policy for repeat auto-rename input while a file op is active.
-    pub(crate) fn handle_busy_browser_auto_rename_intent(
-        &mut self,
-        key: BrowserRenameIntentKey,
-        pending: PendingBrowserAutoRenameIntent,
-    ) -> BrowserRenameBusyDecision {
-        let Some(active) = self.active_browser_rename_intent.as_ref() else {
-            return BrowserRenameBusyDecision::UnrelatedFileOp;
-        };
-        if active == &key
-            || self
-                .queued_browser_auto_rename_intent
-                .as_ref()
-                .is_some_and(|queued| queued.key == key)
-        {
-            return BrowserRenameBusyDecision::Collapsed;
-        }
-        self.queued_browser_auto_rename_intent = Some(pending);
-        BrowserRenameBusyDecision::Queued
-    }
-
-    /// Return whether the same browser rename request is already active.
-    pub(crate) fn browser_rename_intent_is_active(&self, key: &BrowserRenameIntentKey) -> bool {
-        self.active_browser_rename_intent.as_ref() == Some(key)
-    }
-
-    /// Record the latest optimistic Loop/One-shot edit for one source/path.
-    pub(crate) fn begin_looped_metadata_intent(
-        &mut self,
-        source_id: &SourceId,
-        relative_path: &std::path::Path,
-    ) -> u64 {
-        self.next_looped_metadata_intent_id =
-            self.next_looped_metadata_intent_id.saturating_add(1).max(1);
-        let intent_id = self.next_looped_metadata_intent_id;
-        self.looped_metadata_intents
-            .insert((source_id.clone(), relative_path.to_path_buf()), intent_id);
-        intent_id
-    }
-
-    /// Return whether a Loop/One-shot completion still owns the current optimistic edit.
-    pub(crate) fn looped_metadata_intent_matches(
-        &self,
-        source_id: &SourceId,
-        relative_path: &std::path::Path,
-        intent_id: u64,
-    ) -> bool {
-        self.looped_metadata_intents
-            .get(&(source_id.clone(), relative_path.to_path_buf()))
-            .copied()
-            == Some(intent_id)
-    }
-
-    /// Clear a completed Loop/One-shot intent only if it is still the current owner.
-    pub(crate) fn finish_looped_metadata_intent(
-        &mut self,
-        source_id: &SourceId,
-        relative_path: &std::path::Path,
-        intent_id: u64,
-    ) {
-        let key = (source_id.clone(), relative_path.to_path_buf());
-        if self.looped_metadata_intents.get(&key).copied() == Some(intent_id) {
-            self.looped_metadata_intents.remove(&key);
-        }
-    }
-
-    /// Follow a successful rename so stale metadata completions check the live path.
-    pub(crate) fn remap_looped_metadata_intent(
-        &mut self,
-        source_id: &SourceId,
-        old_relative: &std::path::Path,
-        new_relative: &std::path::Path,
-    ) {
-        if old_relative == new_relative {
-            return;
-        }
-        let old_key = (source_id.clone(), old_relative.to_path_buf());
-        if let Some(intent_id) = self.looped_metadata_intents.remove(&old_key) {
-            self.looped_metadata_intents
-                .insert((source_id.clone(), new_relative.to_path_buf()), intent_id);
-        }
-    }
-}
-
 /// UI row state for one requested path in an active auto-rename batch.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub(crate) enum AutoRenameBatchRowState {
@@ -432,89 +103,6 @@ struct ActiveAutoRenameBatchState {
     states: HashMap<PathBuf, AutoRenameBatchRowState>,
     remaps: HashMap<PathBuf, PathBuf>,
     current_requested_path: Option<PathBuf>,
-}
-
-impl ActiveAutoRenameBatchState {
-    fn new(source_id: SourceId, paths: Vec<PathBuf>) -> Self {
-        let states = paths
-            .iter()
-            .cloned()
-            .map(|path| (path, AutoRenameBatchRowState::Queued))
-            .collect();
-        Self {
-            source_id,
-            requested_paths: paths,
-            states,
-            remaps: HashMap::new(),
-            current_requested_path: None,
-        }
-    }
-
-    fn apply_progress(&mut self, progress: SampleAutoRenameProgress) {
-        match progress {
-            SampleAutoRenameProgress::Active { old_relative } => {
-                self.current_requested_path = Some(old_relative.clone());
-                self.states
-                    .insert(old_relative, AutoRenameBatchRowState::Active);
-            }
-            SampleAutoRenameProgress::Completed {
-                old_relative,
-                new_relative,
-            } => {
-                self.current_requested_path = None;
-                if old_relative != new_relative {
-                    self.remaps
-                        .insert(old_relative.clone(), new_relative.clone());
-                }
-                self.states
-                    .insert(old_relative, AutoRenameBatchRowState::Completed);
-            }
-            SampleAutoRenameProgress::Skipped { old_relative, .. } => {
-                self.current_requested_path = None;
-                self.states
-                    .insert(old_relative, AutoRenameBatchRowState::Skipped);
-            }
-            SampleAutoRenameProgress::Failed { old_relative, .. } => {
-                self.current_requested_path = None;
-                self.states
-                    .insert(old_relative, AutoRenameBatchRowState::Failed);
-            }
-        }
-    }
-
-    fn snapshot(&self) -> ActiveAutoRenameBatchSnapshot {
-        ActiveAutoRenameBatchSnapshot {
-            source_id: self.source_id.clone(),
-            rows: self
-                .requested_paths
-                .iter()
-                .map(|requested_path| AutoRenameBatchRowSnapshot {
-                    requested_path: requested_path.clone(),
-                    current_path: self
-                        .remaps
-                        .get(requested_path)
-                        .cloned()
-                        .unwrap_or_else(|| requested_path.clone()),
-                    state: self
-                        .states
-                        .get(requested_path)
-                        .copied()
-                        .unwrap_or(AutoRenameBatchRowState::Queued),
-                })
-                .collect(),
-            current_path: self.current_requested_path.as_ref().map(|path| {
-                self.remaps
-                    .get(path)
-                    .cloned()
-                    .unwrap_or_else(|| path.clone())
-            }),
-            remaps: self
-                .remaps
-                .iter()
-                .map(|(old, new)| (old.clone(), new.clone()))
-                .collect(),
-        }
-    }
 }
 
 /// Decision for browser rename input received while the generic file-op lane is busy.
